@@ -2,9 +2,10 @@ name: flutter_simple_shopify_example
 description: A simple shopify example.
 
 publish_to: 'none'
+version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
