@@ -7,19 +7,21 @@ part 'products.g.dart';
 @freezed
 class Products with _$Products {
   const Products._();
-  factory Products(
-      {required List<Product> productList,
-      required bool hasNextPage}) = _Products;
+  factory Products({
+    required List<Product> productList,
+    required bool hasNextPage,
+  }) = _Products;
 
   static Products fromGraphJson(Map<String, dynamic> json) {
     return Products(
-        productList: _getProductList(json),
-        hasNextPage: (json['pageInfo'] ?? const {})['hasNextPage'] ?? false);
+      productList: _getProductList(json),
+      hasNextPage: (json['pageInfo'] ?? const {})['hasNextPage'] ?? false,
+    );
   }
 
   static List<Product> _getProductList(Map<String, dynamic> json) {
     return (json['edges'] as List?)
-            ?.map((e) => Product.fromGraphJson(e ?? const {}))
+            ?.map((e) => Product.fromGraphJson(e?['node'] ?? const {}))
             .toList() ??
         const <Product>[];
   }
